requirements.txt
# Python 3.10+ recommended
aiogram>=3.4,<4.0
SQLAlchemy>=1.4,<2.0
pandas>=2.2
openpyxl>=3.1
PySocks>=1.7
magic-filter>=1.0